--- src/Acmebot.App/Options/IonosCloudOptions.cs
@@ -1,6 +1,0 @@
-﻿namespace Acmebot.App.Options;
-
-public class IonosCloudOptions
-{
-    public required string Token { get; set; }
-}
