+++ src/Acmebot.App/Options/IonosDnsOptions.cs
@@ -1,0 +1,6 @@
+﻿namespace Acmebot.App.Options;
+
+public class IonosDnsOptions
+{
+    public required string ApiKey { get; set; }
+}
